fix(01-02): LRCLIB duration uses Double preserving fractional seconds
- LrcLibApi.getLyrics parameter changed from Int? to Double?
- LrcLibProvider converts durationMs via it/1000.0 instead of (it/1000).toInt()
- Updated existing test assertion from duration=238 to duration=238.0
- 2 new tests verify float duration: 238500ms -> 238.5, 180000ms -> 180.0

+++ musicmeta-core/src/main/kotlin/com/landofoz/musicmeta/provider/lrclib/LrcLibApi.kt
@@ -22,7 +22,7 @@ class LrcLibApi(
         artist: String,
         track: String,
         album: String? = null,
-        durationSec: Int? = null,
+        durationSec: Double? = null,
     ): LrcLibResult? = rateLimiter.execute {
         val url = buildString {
             append("$BASE_URL/api/get?")
